workspace restructured
