+++ src/backend/go.mod
@@ -1,0 +1,17 @@
+module src/backend
+
+go 1.21.4
+
+require (
+	github.com/daviddengcn/go-villa v0.0.0-20200811194146-68107afb6d76
+	github.com/mattn/go-sqlite3 v1.14.18
+	gorm.io/driver/sqlite v1.5.4
+	gorm.io/gorm v1.25.5
+)
+
+require (
+	github.com/golangplus/bytes v1.0.0 // indirect
+	github.com/golangplus/sort v1.0.0 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+)
